requirements.txt
pandas
python-dotenv
jupyter
git+https://github.com/EnergieID/entsoe-py.git